Refactor buffer passing and allocation
Let the buffer in a separate heap space to allow to grow it later. Make
the interface for buffer passing a pointer, just setting data and length
to the correct values. This avoids data copy just as before and besides
allows us to play nice with the buffer separation as explained above.