Split packets on mavlink boundaries
This is done for both UDP and UART. We fill the rx_buf buffer until we
have a valid mavlink packet or we notice the packet won't be valid.
For now we don't check CRC or signature since that's supposed to be done
by the client nonetheless, as if there was no routing in between.

This already prepares the codebase to be able to read sysid and
compid.