Only do MAVLink CRC check for messages received from UART
Messages received from UDP already have checksum to check data
integrity so it is safe presume that corrupt packages will not
be handled on MAVLink router.

There is no much changes in CPU usage with this patch because
QGroundControl don't send a lot of messages but in a setup
with multiple network endpoint it can save some CPU time
in user space.