Allow more than one 'master' endpoint
Currently, we have one 'master' endpoint - which is supposed to be the
aircraft. Master endpoints are those which are always sending messages -
we listen to them and forward to and from them. Other endpoints may have
this property as well, so we must handle them.

mavlink-router now takes more positional arguments that define extra
'master' endpoints. Note that we are still limited to one serial
endpoint - only multiple UDP endpoints (or one serial and multiple UDP)
are handled in this patch.