Add functions to read informations from a .conf file
A really simple API, in fact: only methods are `parse_file`,
`next_section` and  `next_from_section`. However, it is be able
to retrieve multiple keys with same name, and should be simple
to use.

It's a bit heavy on _small_ mallocs: if that proves to be a problem, a
different allocation scheme can be used with minor changes.