tcp: Reconnect to TCP endpoints in the event of a disconnection
When connection to flight stack - be with `-p` argument or `TcpEndpoint`
config, connection may be cut. This patch adds the ability to reconnect
after a timeout, default 5 seconds.

To change timeout, one needs to define endpoint of configuration file,
and use config `RetryTimeout`, as:

[TcpEndpoint baz]
...
RetryTimeout = 10

This sets timeout to 10 seconds. Setting it to 0, disable reconnection
attempts.