log: Fix TOCTOU on log file creation and change log file format
To ensure a unique name was obtained, code checked filename using
`access`, however, this is prone to Time Of Check/Time Of Use bugs.

Also, instead of using current date and time (and eventually a increasing
suffix), uses a increasing prefix and date and time as filename.
For instance, three consecutive runs would yield to:

   000-2017-03-21_11-43-50.bin
   001-2017-03-21_11-43-52.bin
   002-2017-03-21_11-43-53.bin

Note that prefix will increase independent of date and time - it
actually looks for files on current log dir to find out next prefix.
Currently, maximum prefix is 999.