endpoints: Use a property to define if an endpoint is valid
A TcpEndpoint may become 'invalid' if other side disconnects. In this
case, it needs to be removed from current list of TcpEndpoints as well
as from epoll. So far, this has been done indirectly on Mainloop.

However, this brings some problems: from time to time, we need to check
if an endpoint is TcpEndpoint, in order to handle this invalidation
thing. Sometimes a dynamic_cast can be used, but usually we avoid it,
with some solution that may not add to code readability.

This patch introduces a 'valid' (is_valid()) property to Pollable class:
this way, and endpoint can mark itself invalid, so Mainloop can take
proper measure.

Although being used currently only on TcpEndpoints, this could be
extended to unify Timeout (another Pollable) 'remove_me' propery as
well.