mainloop: Rename struct opt so it won't be shadowed by element opt
Structs in C++ can be accessed with or without the struct keyword, so
this is not a problem for most cases. But in order to use sizeof with
an specific field it is necessary to access it without using the struct
keyword.

Renaming from struct opt to struct options in order to be able to use
sizeof(options::endpoints), for example.

Signed-off-by: Otavio Pontes <otavio.pontes@intel.com>