endpoint: `read_msg` return if a message is certainly valid or not
`Endpoint::read_msg` method returned `1` for any sucessful reading,
disregarding if message is certainly valid or not. Now, it returns
`ReadOk` for certainly valid messages, and `ReadUnidentified` otherwise.

A message is considered `certainly valid` if it has a known message id.
Otherwise, it may be an id shared by endpoints but uknown to
mavlink-router, or just garbage.

This new return value will help on situations where we expect only valid
messages.