Uart Endpoint: Accept a comma separated list of baudrates
Sometimes user may not know (or not have a way to know) correct baudrate
for an UART device. mavlink-router now accepts a comma separated list of
baudrates. In this case, if no correct readings are performed on first
baudrate, it'll try again on next of the list. It will repeat this
behaviour (even circling the list) until it gets a correct reading on a
given baud.

A correct reading is one with valid message id, identified by
ReadOk.

Note that this list behaviour is only available for baudrates defined on
config files - it's not possible use it on argument values, as in
`mavlink-routerd /dev/ttyS0:300,600`. But it perfectly fine on config
file:

[UartEndpoint]
Device = /dev/ttyS0
Baud = 300,600