endpoint: reset uart to raw mode when opening
When opening a UART, make it raw and set a speed known to work on most
hardware.

We need to add a reset_uart() function in a different
compilation unit because depending on the libc it's not possible to
include termios.h and asm/termbits.h at the same time. We need the
latter in order to use termios2 in UartEndpoint.