endpoint: reset uart to raw on exit
Some programs are buggy and don't set vmin/vtim to whatever they need.
It should be their duty to set the UART parameters, but we can at least
play nice and reset it to "raw" and speed of 1200, which should make
most of other programs to want to reconfigure the UART when starting.