Replace log_error_errno
The first argument was actually used to tell log_error_errno() to return
a negative value (usually -errno). This was to allow logging and
returning an error at the same time. It doesn't make sense if we are not
chaining the return.

Since it was not a very used feature, this is removed and the few places
in which it was, a separate return was used.