code shuffle: move source files to src directory
Instead of having all source files on top dir, move them to src,
respecting the following: common files, potentially shareable with
another binaries on `src/common` and those exclusive of mavlink-routerd
itself on `src/mavlink-router`.

One may suspect that `mainloop` and `pollable` classes belong to
`src/common` - and they are right. However, currently, `mainloop` is the
one class that perform the routing, so it will live on
`src/mavlink-router` till it's converted to a pure mainloop.