examples: Add a simple example of a arm authorizer
This is example of a arm authorizer, before the vehicle actually arm
and fly it will request authorization, so authorizer(with more
intelligence) can check for example if the aero space is available or
if the weather is good and authorize or not the flight.

Right now it only authorizes if the altitude of all the waypoints of
the mission loaded is <= 15 meters.