log: Fix buffer overflow
MAVLink 2.0 messages can have extensions which allows new fields
to be added. This happened to mavlink_command_ack_t, gaining an
additional byte with the percentage of progress in case a command
takes a long time to finish.

So if MAVLink router was built with a MAVLink version without
this extension and the flight stack was built with, it would case
a buffer overflow on MAVLink router.

According to the spec, the rules for extensions are:

	(...) the fields after the extensions line are extended fields. The rules
	for extended messages are:

	* if sent as a MAVLink1 message then extended fields are not sent
	* if received by an implementation that doesn't have the extended
	  fields then the fields will not be seen
	* if sent by an implementation that doesn't have the extended fields
	  then the recipient will see zero values for the extended fields

So here what we do is to truncate the message to be considered having
the size of the known message rather than the payload length.