Fix invalid read when parsing UART baudrate from command line
Now the last parameter of add_uart_endpoint() is a string with all
the baudrates that it should try to use but if uses did not provided
a baudrate in the commmand line it would call add_uart_endpoint()
with a invalid string pointer.

This will use the number that split_on_colon() had parsed if it
is a valid number use the string if not call add_uart_endpoint()
with a null pointer and it will use the default baudrate.