endpoint: don't send msg back to the same endpoint
If an endpoint is sending a packet through a channel and there are more
components IDs on the same endpoint, mavlink-router would send back any
message through the same channel.  It's never the expected behavior to
receive the message reflected back and there's no way to the other end
to deal with it. Exception would be of course check the destination
sysid/compid, but those may not be present.

This also changes the behavior if the same vehicle/component is
connected through multiple endpoints on mavlink-router, each of them
having more components. mavlink-router will now drop the packet do avoid
loops.

This is an alternative to the fix provided in
https://github.com/01org/mavlink-router/pull/113 and should fix the
issue in https://github.com/01org/mavlink-router/issues/107.

It's normal to have more components connected to the same endpoint.
What's not normal is that these components depend on mavlink-router to
talk between themselves.

Examples:

1)	GCS -> FC   -> mavlink-router -> App
	       FC -/

2)	GCS -> FC   -> mavlink-router -> mavros --> App1
	       FC -/                            \-> App2

In the first case, the flight stack needs to check if it should handle
the message or if it should forward it. Both ArduPilot and PX4 already
do that.

On the second case, App1 and App2 should not be talking to each other
via mavlink-router. They should either
	a) be communicating directing (alas they are ROS applications
	and should keep the mavlink serialization and de-serialization
	out of the game) or

	b) make mavros route the packet. There's no way for
	mavlink-router to do the right thing if it sends a (possibly
	broadcast message) over to the same endpoint that sent it.