mainloop: handle epoll_wait error
epoll can fail for example when an UART FTDI cable is plugged out and the
associated device becomes invalid. In that case we remove the file
descriptor from the list.

We also quit mavlink-router, because in most cases the error is fatal,
and an external component (e.g. systemd) can restart it.