Fix undefined behavior in mainloop
C++11 standard section 1.9: the value of any object not in either of these two
categories [std::atomic or volatile sig_atomic_t] that is modified by the
[signal] handler becomes undefined.

Replace "volatile bool" with atomic<bool> which is safe according to standard.
Also, wrap this in request_exit method that can be called externally.