Timeout timer - change to NONBLOCK otherwise we could hang in a timer read operation
This could happen if in the same ground of events in the mainloop we have
and operation disarming the timer followed by a timer read. This is a
corner case that can happen more often when using mavlink message
coalescing since the timer is continuously used. When this issue occurs
the entire mavlink router mainloop is hanging since the timer read is
done in this loop.