Do not exit on TCP disconnects
TCP connections are expected to be dynamic. Just like we accept new
connection on the fly, we should also gracefully accept disconnects.
This was the behavior before, but change when we decided to handle UART
disconnections as fatal in 7e7ea939c8d7 ("mainloop: handle epoll_wait error").
For UART, this should not be common as the OS is simply removing the
device under us due to physical disconnection. This restores the previous
behavior for TCP.

In a poll loop it's usual to handle the POLLERR before, but in this case
we leave it after the POLLIN/POLLOUT. Even if we have a POLLERR
we allow the read()/write() calls to be done, so to give the endpoint
the chance to set the `is_valid()` flag.

While at it, fix minor coding style on indentation.

Fix: #252