build: show if tests are enabled
It's bad that we automatically enable/disable if something is available
or not when building mavlink-router. But it's even worse that we don't
even show if something is left out. For tests it's misleading because
`make check` will just pass if we don't have gtest available (a main
requirement). In the past we had gtest as a submodule so we didn't have
to worry about it, but now we only use the system-provided gtest.