Do not warn on TIOCSSERIAL if it's expected
Very few drivers in Linux implement ASYNC_LOW_LATENCY. And a few drivers
don't even have the set_serial() hook function set. When there is no
set_serial() the driver returns ENOTTY and when the flag is not
supported, it is supposed to return ENODEV. Only warn if we get a
different error.