binlog: Do not ACK then NACK the same packet
An off by one error in the for loop for sending NACK's on the unseen
seqno's cause us to send a NACK for the packet. Last acked was the
packet we just sent an ACK for, so we don't want to then send a NACK out
for it.

The other change this does is just swap to using the MAVLink enum's for
ACK/NACK for code readibility