build: move src/mavlink-router/* to src/
No need to keep an extra dir. We already don't have a clear separation
on what is exclusive to mavlink-router to what is "library" / "support"
(example:  the mainloop implementation. Move it one directory down
and once we start a proper split we can move code to common/ (eventually
renaming it to something better).