logendpoint: move logging start timeout
Once upon a time there was just a single timeout, so it was easy to
track. Now with more timeouts it's becoming difficult to track what is
the handle and callback for each of them. Rename _start_timeout()
to _logging_start_timeout() and move the handle to a _timeout sub
struct to avoid name clash. Eventually we will move the other timeouts
there too.