Use pre-built mavlink library
Generating the mavlink C library on the fly is good because it allows us
to experiment new changes while also change the source code. However
this benefit is not used much for mavlink-router - one of the reasons
because the messages are probably developed first with the flight stack
and GCS code.

One of the reasons to migrate out of generating the headers is that this
will make it easier to migrate to meson, as it is currently not
compatible with generating the mavlink header files.

Changes to CI to make it pass:
  - Use actions/checkout@v2 that works better with submodules
  - Update alpine to the latest release
  - Install tools before doing checkout since we want latest tools
    installed, like git
  - Do not require python anymore