ci: migrate to meson
dockcross was not doing much for us, so ditch it and instead just use
the builds with the cross toolchain installed and rootfs prepared.  This
also drops old distro versions to reduce maintenance burden.

Also let's drop debian packaging for now while we replace the build
system. This can return later if needed and should be easier now with
the sysroot integration.

We need to bump the minimum meson version supported due to the use of
constants. In theory this got added in 0.55, but I could only test on
0.57+, so let's make it 0.57.