Add dockerfile and script to create sysroot
There are several tools to create a sysroot/rootfs that could be used. A
simple one can be accomplished by using a docker container, prepared for
the target architecture. We leverage binfmt_misc to be able to run
foreign binaries and thus being able to install additional packages and
possibly configuration in the container.