endpoint: add common header fields to struct buffer
We ended up adding a lot of fields in the stack in order to route
messages and even parsing the message again on logendpoint just because
we need more fields to take some decisions.

Instead of doing that, start filling out new fields in struct buffer
that have the info about the (first) message the data field points to.
When we detect we have a valid message in Endpoint::read_msg() we copy
the relevant fields so they can be used by other parts of the code.

Once other parts of the code are adapted we can stop passing those
fields as argument.