Fix assert crash when stopping log
When we exit mavlink-router we call logger->stop() to stop
and properly finalize the flight stack log. However this
call may want to send something to the flight stack to stop
sending the log.

We don't care about that call as if we are really stopping the flight
stack should detect the other side died/disconnected after a timeout.
Instead of marking the mainloop as not initiliazed upon signal,
serialize it with the other shutdown task in the main loop.

This fixes the crashes:
    Logging target system_id=1 on 00001-2021-12-29_23-57-32.bin
    mavlink-routerd: ../src/mainloop.h:77: static Mainloop&Mainloop::get_instance(): Assertion `_initialized' failed.
    Aborted

    and

    Opened TCP Client [4]CLI: 127.0.0.1:5760
    Logging target system_id=1 on 00054-2022-01-02_23-40-14.bin
    ^CTCP [4]CLI: Connection closed

Also, when shutting down we are not gaining much by keeping that flag
around. The possible objects using the mainloop will be destroyed before
the mainloop itself. So quick fix is to just remove the setting back to
false. A possibly better thing would be to make sure all objects are
properly terminated, but it's not needed for this issue.