ci: workaround broken gtest.pc in ubuntu
The gtest.pc in Ubuntu 20.04 has the includedir relative to the pc file,
but when we have multilib, the pc file is installed to
i386-linux-gnu/pkgconfig, making the final dir be /usr/lib/include.

There is already a bug report about that:
https://bugs.launchpad.net/ubuntu/+source/googletest/+bug/1876042
And it's also fixed in later Ubuntu versions. For CI purposes, let's
just wrap the pkg-config binary, hard-coding the value of the prefix
variable to /usr instead of relative to the pc file.

While at it, move the cross-compiler definitions to a
meson-cross-i686.ini file and replace m32 with i686 for consistency.