README: Write down some guidelines
These guidelines were adapted from https://github.com/rostedt/libccli.
For people like me who generally work in the Linux kernel, it's hard to
apply some practices common to other projects. This is not as strict as
rules to contribute to Linux kernel, but at least give a hint on how
to make it easier for the PR to be reviewed and applied.