TCP endpoint: Reject message when not connected
When a TCP endpoint has lost it's connection to the server, it'll try to
reconnect. During this time, the socket it closed, so no writes can happen. But
this will only be noted during the write_msg method and lead to an error
message, which isn't relevant since it's expected, that no write can happen on
a TCP endpoint which tries to reconnect.