TCP endpoint: Ignore write error during reconnect attempt
When a TCP endpoint detects an error, it schedules a reconnect (if not
disabled). During the reconnection, the fd is invalid, but the endpoint
still exists, so the write_msg() method is still called. But write_msg
gives an error when the fd is invalid, even though this is expected
as long as the reconnect attempts are running.