feat: On CRC errors keep routing the messages
This treats messages with mismatching CRC
the same as unknown messages and improves
consistency accross mavlink-router.

If a CRC error is encountered, it is either due
to corruption or changed message definition of the
specific message id. In this case, it is safe to just
treat it as an unknown message and route it anyways.

ULOG and Binary endpoints however still reject messages
with a mismatching CRC, since these endpoints are the
final consumer of the mavlink messages. This makes
sure that only correct and valid data is logged to file.

Update src/mavlink-router/endpoint.cpp

Co-authored-by: Alessandro Simovic <alessandro@auterion.com>
Co-authored-by: Beat Küng <beat-kueng@gmx.net>