Do not reset uart
This fixes the issue exhibited in #213. I have tested this on a Jetson
TX2 and confirmed that by not performing the reset, the UART port is
still accessible with minicom after shutting down mavlink router. For
what its worth, setting stty's sane on the UART does not cause the same
problem, so there is evidently some difference between the reset_uart
and stty sane despite the comment in the code.

The UART reset was added in commit 4b719070b178 ("endpoint: reset uart
to raw on exit") because we use vmin/vtim in the UART to be notified of
new data and some programs weren't working after shutting down
mavlink-router. It should be safe to remove this by now.